+++ packages/figma/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @seed-design/figma
 
+## 1.3.9
+
+### Patch Changes
+
+- 72dfd8f: Figma Codegen 및 MCP에서 [Slot 레이어](https://help.figma.com/hc/en-us/articles/38231200344599-Use-slots-to-build-flexible-components-in-Figma)을 지원하도록 업데이트합니다.
+- Updated dependencies [846036b]
+- Updated dependencies [e78d460]
+  - @seed-design/css@1.2.7
+
+## 1.3.8
+
+### Patch Changes
+
+- 7cbe9f3: Figma 레이어 normalizer 과정에서, 일부 텍스트 레이어에서 `fontFamily` 관련 크래시가 발생할 수 있는 문제를 수정합니다.
+
+## 1.3.7
+
+### Patch Changes
+
+- 534eb1c: 신규 컴포넌트 `ContentPlaceholder`에 대한 Figma Codegen을 지원합니다. `ImageFrame` codegen 시 `fallback` 값으로 `ContentPlaceholder` 코드를 생성합니다.
+- Updated dependencies [4f56e76]
+- Updated dependencies [32938de]
+- Updated dependencies [67a7780]
+- Updated dependencies [3273312]
+- Updated dependencies [9d9b891]
+  - @seed-design/css@1.2.6
+
 ## 1.3.6
 
 ### Patch Changes